refactor(city): extract shared city projection service (#1024)
